customer seller get, register, update (incorrect)

+++ domain/model/user/admin.go
@@ -1,6 +1,6 @@
 package model
 
 type Admin struct {
-	UserID int  `form:"user_id" json:"user_id"`
-	User   User `form:"user" json:"user"`
+	UserID int64 `form:"user_id" json:"user_id"`
+	User   User  `form:"user" json:"user"`
 }
