+++ domain/model/user/ticket.go
@@ -1,10 +1,10 @@
 package model
 
 type Ticket struct {
-	ID       int    `form:"id" json:"id"`
+	ID       int64  `form:"id" json:"id"`
 	Category string `form:"category" json:"category"`
 	Inquiry  string `form:"inquiry" json:"inquiry"`
-	Reply	 string `form:"reply" json:"reply"`
+	Reply    string `form:"reply" json:"reply"`
 	UserID   int    `form:"user_id" json:"user_id"`
 	User
 }
